refactor: class명 변경
